ユーザ認証テストを自動テストからコマンド形式に変更
